Cleanup; component for search overlay
